modified:   doc/Data/FileInfo.nd 	modified:   doc/Data/IndexInfo.nd 	modified:   doc/Data/SymbolTable.nd 	modified:   doc/files/utils/src/svcs_primitives-h.html 	modified:   doc/index/Functions.html 	modified:   doc/index/General.html 	deleted:    utils/src/stat_cov_cocket_utils.h 	modified:   utils/src/svcs_primitives.c 	modified:   utils/src/svcs_primitives.h

+++ utils/src/svcs_primitives.h
@@ -50,7 +50,7 @@ Data Types: integer,  double
   Returns: 
   hash value
 */
-unsigned long svcs_prim_hash(const char *str);
+double svcs_prim_hash(const char *str);
 
 /*
   Function: svcs_prim_error
